Extract lens dimensions from lens editor app
