+++ x-pack/platform/plugins/shared/lens/test/scout/common/ui/fixtures/page_objects/lens/lens_editor_helpers.ts
@@ -5,6 +5,9 @@
  * 2.0.
  */
 
+/** `useDebouncedValue` waits 256ms before committing; add margin for a busy main thread. */
+export const FORMAT_PARAM_DEBOUNCE_FLUSH_MS = 500;
+
 /** Default timeout for `page.waitForFunction` readiness waits. */
 export const WAIT_FOR_FUNCTION_TIMEOUT_MS = 10_000;
