+++ x-pack/platform/plugins/shared/lens/test/scout/core/ui/parallel_tests/drag_and_drop/field_to_workspace.spec.ts
@@ -29,7 +29,7 @@ spaceTest.describe(
       const { lens } = pageObjects;
 
       await lens.dragFieldToWorkspace('machine.os.raw', testData.XY_CHART);
-      await expect(lens.getDimensionTriggersLocator('lnsXY_xDimensionPanel')).toHaveText(
+      await expect(lens.dimensions.getDimensionTriggersLocator('lnsXY_xDimensionPanel')).toHaveText(
         'Top 9 values of machine.os.raw'
       );
       expect(await lens.getChartSwitchType()).toBe('Bar');
@@ -39,7 +39,7 @@ spaceTest.describe(
       const { lens } = pageObjects;
 
       await lens.dragFieldToWorkspace('@timestamp', testData.XY_CHART);
-      await expect(lens.getDimensionTriggersLocator('lnsXY_xDimensionPanel')).toHaveText(
+      await expect(lens.dimensions.getDimensionTriggersLocator('lnsXY_xDimensionPanel')).toHaveText(
         '@timestamp'
       );
       expect(await lens.getChartSwitchType()).toBe('Line');
@@ -55,11 +55,13 @@ spaceTest.describe(
           async () => {
             await lens.dragFieldToWorkspace('@timestamp', testData.XY_CHART);
             await lens.dragFieldToWorkspace('clientip', testData.XY_CHART);
-            await expect(lens.getDimensionTriggersLocator('lnsXY_splitDimensionPanel')).toHaveText([
-              'Top 9 values of clientip',
-            ]);
-            await lens.openDimensionEditor('lnsXY_splitDimensionPanel > lns-dimensionTrigger');
-            expect(await lens.isTopLevelAggregation()).toBe(true);
+            await expect(
+              lens.dimensions.getDimensionTriggersLocator('lnsXY_splitDimensionPanel')
+            ).toHaveText(['Top 9 values of clientip']);
+            await lens.dimensions.openDimensionEditor(
+              'lnsXY_splitDimensionPanel > lns-dimensionTrigger'
+            );
+            expect(await lens.dimensions.isTopLevelAggregation()).toBe(true);
             await lens.closeDimensionEditor();
           }
         );
@@ -71,9 +73,9 @@ spaceTest.describe(
             await lens.dragFieldToWorkspace('utc_time', testData.XY_CHART);
             await lens.searchField('client');
             await lens.dragFieldToWorkspace('clientip', testData.XY_CHART);
-            await expect(lens.getDimensionTriggersLocator('lnsXY_xDimensionPanel')).toHaveText([
-              'utc_time',
-            ]);
+            await expect(
+              lens.dimensions.getDimensionTriggersLocator('lnsXY_xDimensionPanel')
+            ).toHaveText(['utc_time']);
           }
         );
       }
