+++ x-pack/platform/plugins/shared/lens/test/scout/core/ui/parallel_tests/tagcloud_filter.spec.ts
@@ -30,7 +30,7 @@ spaceTest.describe('Lens tag cloud filter', { tag: '@local-stateful-classic' }, 
         field: 'ip',
         keepOpen: true,
       });
-      await lens.setTermsNumberOfValues(5);
+      await lens.dimensions.setTermsNumberOfValues(5);
       await lens.closeDimensionEditor();
 
       await lens.configureDimension({
