+++ x-pack/platform/plugins/shared/lens/test/scout/open_in_lens/ui/parallel_tests/agg_based/gauge.spec.ts
@@ -36,8 +36,8 @@ spaceTest.describe('Lens open in Lens — agg-based Gauge', { tag: tags.deployme
     await lens.waitForVisualization('gaugeChart');
     expect(await lens.layers.getLayerCount()).toBe(1);
 
-    await expect(lens.dimensionTriggerLocator).toHaveCount(3);
-    const dimensions = await lens.getDimensionTriggers();
+    await expect(lens.dimensions.dimensionTriggerLocator).toHaveCount(3);
+    const dimensions = await lens.dimensions.getDimensionTriggers();
     await expect(dimensions[0]).toHaveText('Count');
 
     const { bullet } = await lens.getCurrentChartDebugState('gaugeChart');
@@ -53,8 +53,8 @@ spaceTest.describe('Lens open in Lens — agg-based Gauge', { tag: tags.deployme
     await lens.waitForVisualization('gaugeChart');
     expect(await lens.layers.getLayerCount()).toBe(1);
 
-    await expect(lens.dimensionTriggerLocator).toHaveCount(3);
-    const dimensions = await lens.getDimensionTriggers();
+    await expect(lens.dimensions.dimensionTriggerLocator).toHaveCount(3);
+    const dimensions = await lens.dimensions.getDimensionTriggers();
     await expect(dimensions[0]).toHaveText('Average machine.ram');
     await expect(dimensions[1]).toHaveText('Static value: 0');
     await expect(dimensions[2]).toHaveText('Static value: 100');
@@ -84,8 +84,8 @@ spaceTest.describe('Lens open in Lens — agg-based Gauge', { tag: tags.deployme
     await lens.waitForVisualization('gaugeChart');
     expect(await lens.layers.getLayerCount()).toBe(1);
 
-    await expect(lens.dimensionTriggerLocator).toHaveCount(3);
-    const dimensions = await lens.getDimensionTriggers();
+    await expect(lens.dimensions.dimensionTriggerLocator).toHaveCount(3);
+    const dimensions = await lens.dimensions.getDimensionTriggers();
     await expect(dimensions[0]).toHaveText('Average machine.ram');
     await expect(dimensions[1]).toHaveText('Static value: 0');
     await expect(dimensions[2]).toHaveText('Static value: 15000000000');
