+++ x-pack/platform/plugins/shared/lens/test/scout/open_in_lens/ui/parallel_tests/agg_based/metric.spec.ts
@@ -33,7 +33,7 @@ spaceTest.describe('Lens open in Lens — agg-based Metric', { tag: tags.deploym
     await convertToLensByTitle({ dashboard }, 'Metric - Basic');
     await lens.waitForVisualization('mtrVis');
 
-    await lens.hoverOverDimensionButton();
+    await lens.dimensions.hoverOverDimensionButton();
     const data = await lens.getMetricVisualizationData();
     expect(data).toHaveLength(1);
     expect(data).toStrictEqual([
@@ -57,11 +57,11 @@ spaceTest.describe('Lens open in Lens — agg-based Metric', { tag: tags.deploym
     await lens.waitForVisualization('mtrVis');
     expect(await lens.layers.getLayerCount()).toBe(1);
 
-    await expect(lens.dimensionTriggerLocator).toHaveCount(1);
-    const dimensions = await lens.getDimensionTriggers();
+    await expect(lens.dimensions.dimensionTriggerLocator).toHaveCount(1);
+    const dimensions = await lens.dimensions.getDimensionTriggers();
     await expect(dimensions[0]).toHaveText('Average machine.ram');
 
-    await lens.hoverOverDimensionButton();
+    await lens.dimensions.hoverOverDimensionButton();
     const data = await lens.getMetricVisualizationData();
     expect(data).toHaveLength(1);
     expect(data).toStrictEqual([
@@ -85,12 +85,12 @@ spaceTest.describe('Lens open in Lens — agg-based Metric', { tag: tags.deploym
     await lens.waitForVisualization('mtrVis');
     expect(await lens.layers.getLayerCount()).toBe(1);
 
-    await expect(lens.dimensionTriggerLocator).toHaveCount(2);
-    const dimensions = await lens.getDimensionTriggers();
+    await expect(lens.dimensions.dimensionTriggerLocator).toHaveCount(2);
+    const dimensions = await lens.dimensions.getDimensionTriggers();
     await expect(dimensions[0]).toHaveText('Overall Max of Count');
     await expect(dimensions[1]).toHaveText('@timestamp');
 
-    await lens.hoverOverDimensionButton();
+    await lens.dimensions.hoverOverDimensionButton();
     const data = await lens.getMetricVisualizationData();
     expect(data).toHaveLength(1);
     expect(data).toStrictEqual([
@@ -124,12 +124,12 @@ spaceTest.describe('Lens open in Lens — agg-based Metric', { tag: tags.deploym
     await lens.waitForVisualization('mtrVis');
     expect(await lens.layers.getLayerCount()).toBe(1);
 
-    await expect(lens.dimensionTriggerLocator).toHaveCount(2);
-    const dimensions = await lens.getDimensionTriggers();
+    await expect(lens.dimensions.dimensionTriggerLocator).toHaveCount(2);
+    const dimensions = await lens.dimensions.getDimensionTriggers();
     await expect(dimensions[0]).toHaveText('Average machine.ram');
     await expect(dimensions[1]).toHaveText('machine.os.raw: Descending');
 
-    await lens.hoverOverDimensionButton();
+    await lens.dimensions.hoverOverDimensionButton();
     await expect
       .poll(async () => lens.getMetricVisualizationData(), { timeout: 20_000 })
       .toStrictEqual([
