+++ x-pack/platform/plugins/shared/lens/test/scout/open_in_lens/ui/parallel_tests/agg_based/table.spec.ts
@@ -48,8 +48,8 @@ spaceTest.describe('Lens open in Lens — agg-based Table', { tag: tags.deployme
     await lens.waitForVisualization('lnsDataTable');
     expect(await lens.layers.getLayerCount()).toBe(1);
 
-    await expect(lens.dimensionTriggerLocator).toHaveCount(1);
-    const dimensions = await lens.getDimensionTriggers();
+    await expect(lens.dimensions.dimensionTriggerLocator).toHaveCount(1);
+    const dimensions = await lens.dimensions.getDimensionTriggers();
     await expect(dimensions[0]).toHaveText('Average machine.ram');
   });
 
@@ -60,11 +60,11 @@ spaceTest.describe('Lens open in Lens — agg-based Table', { tag: tags.deployme
     await lens.waitForVisualization('lnsDataTable');
     expect(await lens.layers.getLayerCount()).toBe(1);
 
-    await expect(lens.dimensionTriggerLocator).toHaveCount(1);
-    const dimensions = await lens.getDimensionTriggers();
+    await expect(lens.dimensions.dimensionTriggerLocator).toHaveCount(1);
+    const dimensions = await lens.dimensions.getDimensionTriggers();
     await expect(dimensions[0]).toHaveText('Average machine.ram');
 
-    await lens.openDimensionEditor('lnsDatatable_metrics > lns-dimensionTrigger');
+    await lens.dimensions.openDimensionEditor('lnsDatatable_metrics > lns-dimensionTrigger');
     expect(
       await getDimensionFlyoutComboBoxSelectedOptions(page, 'lnsDatatable_summaryrow_function')
     ).toStrictEqual(['Sum']);
@@ -77,10 +77,10 @@ spaceTest.describe('Lens open in Lens — agg-based Table', { tag: tags.deployme
     await lens.waitForVisualization('lnsDataTable');
     expect(await lens.layers.getLayerCount()).toBe(1);
 
-    await expect(lens.dimensionTriggerLocator).toHaveCount(2);
+    await expect(lens.dimensions.dimensionTriggerLocator).toHaveCount(2);
 
-    const metricText = await lens.getDimensionTriggerText('lnsDatatable_metrics', 0);
-    const splitRowText = await lens.getDimensionTriggerText('lnsDatatable_rows', 0);
+    const metricText = await lens.dimensions.getDimensionTriggerText('lnsDatatable_metrics', 0);
+    const splitRowText = await lens.dimensions.getDimensionTriggerText('lnsDatatable_rows', 0);
     expect(metricText).toBe('Overall Max of Count');
     expect(splitRowText).toBe('@timestamp');
   });
@@ -92,10 +92,10 @@ spaceTest.describe('Lens open in Lens — agg-based Table', { tag: tags.deployme
     await lens.waitForVisualization('lnsDataTable');
     expect(await lens.layers.getLayerCount()).toBe(1);
 
-    await expect(lens.dimensionTriggerLocator).toHaveCount(2);
+    await expect(lens.dimensions.dimensionTriggerLocator).toHaveCount(2);
 
-    const metricText = await lens.getDimensionTriggerText('lnsDatatable_metrics', 0);
-    const splitRowText = await lens.getDimensionTriggerText('lnsDatatable_rows', 0);
+    const metricText = await lens.dimensions.getDimensionTriggerText('lnsDatatable_metrics', 0);
+    const splitRowText = await lens.dimensions.getDimensionTriggerText('lnsDatatable_rows', 0);
     expect(metricText).toBe('Cumulative Sum of Count');
     expect(splitRowText).toBe('@timestamp');
   });
@@ -109,11 +109,11 @@ spaceTest.describe('Lens open in Lens — agg-based Table', { tag: tags.deployme
       await lens.waitForVisualization('lnsDataTable');
       expect(await lens.layers.getLayerCount()).toBe(1);
 
-      await expect(lens.dimensionTriggerLocator).toHaveCount(3);
+      await expect(lens.dimensions.dimensionTriggerLocator).toHaveCount(3);
 
-      const metricText = await lens.getDimensionTriggerText('lnsDatatable_metrics', 0);
-      const splitRowText1 = await lens.getDimensionTriggerText('lnsDatatable_rows', 0);
-      const splitRowText2 = await lens.getDimensionTriggerText('lnsDatatable_rows', 1);
+      const metricText = await lens.dimensions.getDimensionTriggerText('lnsDatatable_metrics', 0);
+      const splitRowText1 = await lens.dimensions.getDimensionTriggerText('lnsDatatable_rows', 0);
+      const splitRowText2 = await lens.dimensions.getDimensionTriggerText('lnsDatatable_rows', 1);
       expect(metricText).toBe('Count');
       expect(splitRowText1).toBe('@timestamp');
       expect(splitRowText2).toBe('bytes: Descending');
@@ -127,12 +127,15 @@ spaceTest.describe('Lens open in Lens — agg-based Table', { tag: tags.deployme
     await lens.waitForVisualization('lnsDataTable');
     expect(await lens.layers.getLayerCount()).toBe(1);
 
-    await expect(lens.dimensionTriggerLocator).toHaveCount(2);
+    await expect(lens.dimensions.dimensionTriggerLocator).toHaveCount(2);
 
-    const metricText = await lens.getDimensionTriggerText('lnsDatatable_metrics', 0);
-    const percentageColumnText = await lens.getDimensionTriggerText('lnsDatatable_metrics', 1);
+    const metricText = await lens.dimensions.getDimensionTriggerText('lnsDatatable_metrics', 0);
+    const percentageColumnText = await lens.dimensions.getDimensionTriggerText(
+      'lnsDatatable_metrics',
+      1
+    );
 
-    await lens.openDimensionEditor('lnsDatatable_metrics > lns-dimensionTrigger', 0, 1);
+    await lens.dimensions.openDimensionEditor('lnsDatatable_metrics > lns-dimensionTrigger', 0, 1);
     expect(
       await getDimensionFlyoutComboBoxSelectedOptions(page, 'indexPattern-dimension-format')
     ).toStrictEqual(['Percent']);
