+++ x-pack/platform/plugins/shared/lens/test/scout/open_in_lens/ui/parallel_tests/tsvb/gauge.spec.ts
@@ -92,7 +92,7 @@ spaceTest.describe('TSVB Gauge - Open in Lens', { tag: tags.deploymentAgnostic }
       { stop: '', color: undefined },
     ]);
     await lens.closePalettePanelFlyout();
-    await lens.closeDimensionEditorPanel();
+    await lens.dimensions.closeDimensionEditorPanel();
   });
 
   spaceTest(
